added like button counter, fixed renaming
