+++ front-end/src/components/PostComponent/PostComponent.css
@@ -46,6 +46,11 @@
   margin-bottom: 8px; /* Space below actions */
 }
 
+.like-button{
+  background: white;
+  color: black;
+}
+
 .post-actions span {
   cursor: pointer;
   transition: transform 0.1s ease-in-out; /* Add subtle interaction feedback */
