+++ front-end/src/components/PostComponent/PostComponent.tsx
@@ -1,6 +1,6 @@
-import React from "react";
+import React, { useState } from "react";
 import "./PostComponent.css";
-import { Post as PostModel } from '../../models/Post';
+import { Post as PostModel } from "../../models/Post";
 
 interface PostComponentProps {
   post: PostModel;
@@ -8,6 +8,19 @@ interface PostComponentProps {
 
 const PostComponent: React.FC<PostComponentProps> = ({ post }) => {
   const { username, profilePicPath: avatar, imagePath: image, description } = post;
+  
+  // State for likes
+  const [likes, setLikes] = useState(0);
+  const [isLiked, setIsLiked] = useState(false);
+
+  const handleLike = () => {
+    if (isLiked) {
+      setLikes(likes - 1);
+    } else {
+      setLikes(likes + 1);
+    }
+    setIsLiked(!isLiked);
+  };
 
   return (
     <div className="post">
@@ -18,12 +31,16 @@ const PostComponent: React.FC<PostComponentProps> = ({ post }) => {
       <img className="post-image" src={image} alt={`Post by ${username}`} />
       <div className="post-content">
         <div className="post-actions">
-          <span className="like" role="button" aria-label="Like post">❤️</span>
+          <button 
+            className={`like-button ${isLiked ? "liked" : ""}`} 
+            onClick={handleLike}
+            aria-label="Like post"
+          >
+            ❤️ {likes}
+          </button>
           <span className="comment" role="button" aria-label="Comment on post">💬</span>
         </div>
-        <p className="post-description">
-          {description}
-        </p>
+        <p className="post-description">{description}</p>
       </div>
     </div>
   );
